+++ boardgamegeek/collection.py
@@ -1,7 +1,7 @@
 from __future__ import unicode_literals
 from copy import copy
-from collections import Iterator
 
+from .exceptions import BoardGameGeekError
 from .games import CollectionBoardGame
 from .utils import DictObject
 
@@ -10,9 +10,21 @@ class Collection(DictObject):
 
     def __init__(self, data):
         kw = copy(data)
+
         if "items" not in kw:
             kw["items"] = []
+
+        self._items = []
         self.__game_ids = set()
+
+        for i in kw["items"]:
+            try:
+                if i["id"] not in self.__game_ids:
+                    self.__game_ids.add(i["id"])
+                    self._items.append(CollectionBoardGame(i))
+            except KeyError:
+                raise BoardGameGeekError("invalid game data")
+
         super(Collection, self).__init__(kw)
 
     def _format(self, log):
@@ -25,11 +37,15 @@ class Collection(DictObject):
             i._format(log)
             log.info("---------------------")
 
-    def _add_game(self, game):
-        # Collections can have duplicate elements (different collection ids), so don't add the same thing multiple times
-        if game["id"] not in self.__game_ids:
-            self._data["items"].append(game)
-            self.__game_ids.add(game["id"])
+    def add_game(self, game):
+        try:
+            # Collections can have duplicate elements (different collection ids), so don't add the same thing multiple times
+            if game["id"] not in self.__game_ids:
+                self._data["items"].append(game)
+                self.__game_ids.add(game["id"])
+                self._items.append(CollectionBoardGame(game))
+        except KeyError:
+            raise BoardGameGeekError("invalid game data")
 
     def __str__(self):
         return "{}'s collection, {} items".format(self.owner, len(self))
@@ -46,8 +62,7 @@ class Collection(DictObject):
 
     @property
     def items(self):
-        # TODO: create this list when _add_game instead of doing it on each items() call.
-        return [CollectionBoardGame(x) for x in self._data["items"]]
+        return self._items
 
     def __iter__(self):
         for item in self._data["items"]: